Refactor webview packages and improve build setup
- Rename packages/shared to packages/webview-shared
- Use @repo namespace for internal packages
- Move api.ts out of react folder (not React-specific)
- Move tsconfig.webview.json to packages folder for sharing
- Fix tsBuildInfoFile location to prevent stale build issues
- Add concurrently for watch:all command
- Add warnOnUnassignedImports to ESLint for CSS import ordering
- Remove unnecessary visibility handler from TasksPanel
- Add title parameter to getWebviewHtml
- Add index.css placeholder to tasks package
- Rename vite config helper to avoid Vitest detection
