--- packages/shared/src/react/index.ts
@@ -1,2 +1,0 @@
-export { getVsCodeApi, postMessage, getState, setState } from "./api";
-export { useMessage, useVsCodeState } from "./hooks";
