--- packages/tasks/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Coder Tasks</title>
-	</head>
-	<body>
-		<div id="root"></div>
-		<script type="module" src="/src/index.tsx"></script>
-	</body>
-</html>
