+++ packages/tasks/src/index.tsx
@@ -2,6 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
 import App from "./App";
+import "./index.css";
 
 const root = document.getElementById("root");
 if (root) {
