+++ packages/webview-shared/createWebviewConfig.ts
@@ -23,17 +23,19 @@ export function createWebviewConfig(
 			// Skip gzip size calculation for faster builds
 			reportCompressedSize: false,
 			rollupOptions: {
+				// HTML is generated by the extension with CSP headers
+				input: resolve(dirname, "src/index.tsx"),
 				output: {
 					entryFileNames: "index.js",
 					assetFileNames: "index.[ext]",
 				},
 			},
-			// No sourcemaps in production (not needed in packaged extension)
+			// Keeps extension size down; build locally to map stack traces
 			sourcemap: !production,
 		},
 		resolve: {
 			alias: {
-				"@coder/shared": resolve(dirname, "../shared/src"),
+				"@repo/webview-shared": resolve(dirname, "../webview-shared/src"),
 			},
 		},
 	});
