+++ packages/webview-shared/extension.d.ts
