+++ src/webviews/util.ts
@@ -8,6 +8,7 @@ export function getWebviewHtml(
 	webview: vscode.Webview,
 	extensionUri: vscode.Uri,
 	webviewName: string,
+	title: string,
 ): string {
 	const nonce = getNonce();
 	const baseUri = vscode.Uri.joinPath(
@@ -29,6 +30,7 @@ export function getWebviewHtml(
   <meta charset="UTF-8">
   <meta http-equiv="Content-Security-Policy" content="default-src 'none'; script-src 'nonce-${nonce}'; style-src ${webview.cspSource} 'unsafe-inline'; font-src ${webview.cspSource}; img-src ${webview.cspSource} data:;">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>${title}</title>
   <link rel="stylesheet" href="${styleUri.toString()}">
 </head>
 <body>
