+++ test/tsconfig.json
@@ -4,7 +4,7 @@
 		"baseUrl": "..",
 		"paths": {
 			"@/*": ["src/*"],
-			"@coder/shared": ["packages/shared/extension.d.ts"]
+			"@repo/webview-shared": ["packages/webview-shared/extension.d.ts"]
 		}
 	},
 	"include": [".", "../src"]
